Replace old map widget by new one with legacy wrapper.
