+++ src/DetailView/widgets/ContainerWidget.tsx
@@ -21,7 +21,6 @@ const ContainerWidget = (props: any): JSX.Element => {
     notifier,
     user,
     analytics,
-    googleConfig,
     ViewType,
     containerStore,
     widgets,
@@ -41,7 +40,6 @@ const ContainerWidget = (props: any): JSX.Element => {
         notifier,
         user,
         analytics,
-        googleConfig,
         ViewType,
         containerStore,
       })}
