+++ src/FormControls/MapSelect/index.ts
@@ -1,1 +1,3 @@
 export { MapSelect, MapSelectProps } from './MapSelect'
+export { MapSelectLegacy } from './MapSelectLegacy'
+export { Option as MapSelectOption } from './types'
