+++ src/FormControls/index.ts
@@ -1,3 +1,2 @@
-export * from '../cdk/Controls/GroupControl'
 export * from './MapSelect'
 export * from './MapFilteredSelect'
