+++ src/ResourceComposer/index.tsx
@@ -5,7 +5,6 @@ import { ThemeProvider } from '@chakra-ui/core'
 import type { BaseAdmin } from 'admin'
 import type { Provider } from 'admin/providers/interfaces'
 import type { BaseAnalytic } from 'integration/analytics/base'
-import type { GoogleConfig } from '../integration/google'
 
 import { RenderList } from '../ListView/RenderList'
 import { RenderDetail } from '../DetailView/RenderDetail'
@@ -28,14 +27,12 @@ const AdminResource = ({
   provider,
   user,
   analytics,
-  googleConfig,
 }: {
   name: string
   admin: BaseAdmin
   provider: Provider
   user: object
   analytics: BaseAnalytic | undefined
-  googleConfig?: GoogleConfig
 }): JSX.Element => (
   <Switch>
     <Redirect exact strict from={`/${name}`} to={`/${name}/`} />
@@ -43,14 +40,7 @@ const AdminResource = ({
       <RenderList resourceName={name} admin={admin} provider={provider} user={user} analytics={analytics} />
     </Route>
     <Route exact path={`/${name}/:id`}>
-      <RenderDetail
-        resourceName={name}
-        admin={admin}
-        provider={provider}
-        user={user}
-        analytics={analytics}
-        googleConfig={googleConfig}
-      />
+      <RenderDetail resourceName={name} admin={admin} provider={provider} user={user} analytics={analytics} />
     </Route>
   </Switch>
 )
