+++ src/WizardMaster/components/Wizard.tsx
@@ -12,7 +12,6 @@ import { pushAnalytics } from '../../integration/analytics'
 import type { BaseWizard } from '../interfaces'
 import type { BaseAnalytic } from '../../integration/analytics/base'
 import type { Accessor, DetailObject } from '../../typing'
-import type { GoogleConfig } from '../../integration/google'
 import { getAccessorWithDefault } from '../../DetailView/utils/dataAccess'
 
 type WizardProps = {
@@ -24,7 +23,6 @@ type WizardProps = {
   refreshMainDetailObject: Function
   notifier: BaseNotifier
   analytics: BaseAnalytic | undefined
-  googleConfig: GoogleConfig | undefined
   ViewType: string
   user: object
   style?: object
@@ -54,7 +52,6 @@ const Wizard = (props: WizardProps): JSX.Element => {
     refreshMainDetailObject,
     notifier,
     analytics,
-    googleConfig,
     ViewType,
     user,
     resourceName,
@@ -107,7 +104,6 @@ const Wizard = (props: WizardProps): JSX.Element => {
         refreshMainDetailObject={refreshMainDetailObject}
         notifier={notifier}
         analytics={analytics}
-        googleConfig={googleConfig}
         ViewType={ViewType}
         user={user}
         submitChange={submitChange}
