+++ src/WizardMaster/tests/test_components/WizardStepContainer.test.tsx
@@ -14,7 +14,6 @@ const getComponent = (): JSX.Element => {
   const setMainDetailObject = jest.fn()
   const refreshMainDetailObject = jest.fn()
   const analytics = undefined
-  const googleConfig = undefined
   const user = {}
 
   return (
@@ -28,7 +27,6 @@ const getComponent = (): JSX.Element => {
         refreshMainDetailObject={refreshMainDetailObject}
         notifier={testNotifier}
         analytics={analytics}
-        googleConfig={googleConfig}
         ViewType={ViewType}
         user={user}
         show
