--- src/integration/google/types.ts
@@ -1,3 +1,0 @@
-export interface GoogleConfig {
-  apiKey?: string
-}
